+++ src/models/index.ts
@@ -1,8 +1,9 @@
 export interface PopupInstance {
     close: Function;
+    onCloseClick: Function;
 }
 
 export interface PopupProps {
-    onCloseClick?: (...args: any[]) => any;
-    show?: boolean;
+    onCloseClick: (...args: any[]) => any;
+    show: boolean;
 }